add utk

+++ Data/dataset.py
@@ -5,7 +5,7 @@ from torch.utils.data import Dataset
 class Data(Dataset):
     def __init__(self, X, y):
         self.X = torch.from_numpy(X.astype(np.float32))
-        self.y = torch.from_numpy(y.astype(np.float32))
+        self.y = torch.from_numpy(y.astype(np.float32)).long()
         self.len = self.X.shape[0]
 
     def __getitem__(self, index):
